added support for alternative I2C address allowing two sensors on same I2C bus
Added support using a boolean flag when constructing the ADXL343 class to signal the use of the alternative I2C address.

On breakout boards, the I2C address is set to the alternative (of 0x1D) by connecting SDO to Vcc. Also added a link to the sensortest example to illustrate the use.